+++ scripts/daemon/reload.js
@@ -3,7 +3,7 @@
 var gn = require('gracenode');
 var logger = gn.log.create('daemon-reload');
 var lib = require('./utils/lib');
-var talk = require('./utils/talk');
+var Status = require('./utils/status').Status;
 
 module.exports = function (path) {
 	// listener for exceptions
@@ -12,12 +12,12 @@ module.exports = function (path) {
 		gn.exit();
 	});
 	// check for daemon process
-	talk.setup(path, function (isAppRunning) {
-		if (!isAppRunning) {
+	var status = new Status(path);
+	status.setup(function () {
+		if (!status.isRunning) {
 			logger.error(lib.color('daemon process ' + path + ' not running', lib.COLORS.RED));
-			return gn.exit();
+			return status.end();	
 		}
-		// restart daemon
-		talk.reloadApp();
+		status.reload();
 	});
 };